+++ src/Credfeto.ChangeLog.Tests/ChangeLogUpdaterCreateRelease.cs
@@ -78,6 +78,35 @@ Releases that have at least been deployed to staging, BUT NOT necessarily releas
     }
 
     [Fact]
+    public void UnreleasedWithOnlyWhitespaceEntriesDoesNotCreateARelease()
+    {
+        string changeLog =
+            $@"# Changelog
+All notable changes to this project will be documented in this file.
+
+<!--
+Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
+-->
+
+## [Unreleased]
+### Added
+{"   "}
+### Fixed
+### Changed
+### Removed
+### Deployment Changes
+
+<!--
+Releases that have at least been deployed to staging, BUT NOT necessarily released to live.  Changes should be moved from [Unreleased] into here as they are merged into the appropriate release branch
+-->
+## [0.0.0] - Project created";
+
+        Assert.Throws<EmptyChangeLogException>(() =>
+            ChangeLogUpdater.CreateRelease(Parse(changeLog), "1.0.0", true, Language)
+        );
+    }
+
+    [Fact]
     public void CannotCreateAReleaseThatAlreadyExists()
     {
         const string changeLog =
@@ -157,6 +186,62 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ## [Unreleased]
 ### Added
 - Some Content
+### Fixed
+### Changed
+### Removed
+### Deployment Changes
+
+<!--
+Releases that have at least been deployed to staging, BUT NOT necessarily released to live.  Changes should be moved from [Unreleased] into here as they are merged into the appropriate release branch
+-->
+## [0.0.0] - Project created";
+
+        string updated = Serialise(ChangeLogUpdater.CreateRelease(Parse(changeLog), "1.0.0", true, Language));
+
+        const string expected =
+            @"# Changelog
+All notable changes to this project will be documented in this file.
+
+<!--
+Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
+-->
+
+## [Unreleased]
+### Added
+### Fixed
+### Changed
+### Removed
+### Deployment Changes
+
+<!--
+Releases that have at least been deployed to staging, BUT NOT necessarily released to live.  Changes should be moved from [Unreleased] into here as they are merged into the appropriate release branch
+-->
+
+## [1.0.0] - TBD
+### Added
+- Some Content
+
+## [0.0.0] - Project created";
+
+        this._output.WriteLine(updated);
+        Assert.Equal(expected.ToLocalEndLine(), actual: updated);
+    }
+
+    [Fact]
+    public void ChangeLogWithWhitespaceOnlyEntryExcludesItFromRelease()
+    {
+        string changeLog =
+            $@"# Changelog
+All notable changes to this project will be documented in this file.
+
+<!--
+Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
+-->
+
+## [Unreleased]
+### Added
+- Some Content
+{"   "}
 ### Fixed
 ### Changed
 ### Removed